+++ components/sections/SideBar.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="lg:flex lg:justify-end w-full">
+  <aside class="lg:flex lg:justify-end w-full">
     <div class="lg:ml-12">
       <div class="border border-solid border-gray-400 p-8">
         <h3 class="font-extrabold">欢迎关注公众号（Leafage）</h3>
@@ -167,7 +167,7 @@
         </ul>
       </div>
     </div>
-  </div>
+  </aside>
 </template>