+++ pages/posts/detail/_slug.vue
@@ -51,14 +51,14 @@
             </li>
           </ul>
           <h2 class="my-3 md:text-3xl font-extrabold" v-text="data.title"></h2>
-          <div class="w-full h-full my-8">
+          <figure class="w-full h-full my-8">
             <img :src="data.cover" :alt="data.title" class="w-full" />
-          </div>
-          <p
-            class="my-4 leading-relaxed tracking-wide prose-sm md:prose min-w-full"
+          </figure>
+          <div
+            class="prose-sm md:prose min-w-full"
             v-html="rendered"
-          ></p>
-          <div class="bg-gray-200 my-12 p-8">
+          ></div>
+          <div class="bg-gray-200 my-8 p-8">
             <ul
               class="grid grid-flow-row grid-rows-2 grid-cols-1 md:grid-rows-1 md:grid-cols-2 gap-4 text-xs font-bold"
             >
@@ -109,8 +109,9 @@
           </div>
         </article>
         <div class="my-12">
-          <div class="uppercase font-extrabold">
-            <h3>Top Posts</h3>
+          <div class="flex divide-y-2 divide-gray-400 divide-dotted">
+            <h3 class="uppercase font-extrabold">Top Posts</h3>
+            <span class="flex-1 w-full ml-4 mt-3"></span>
           </div>
           <div class="my-6">
             <div