+++ plugins/markdown.js
@@ -37,7 +37,7 @@ const md = new MarkdownIt({
       try {
         return (
           '<pre class="hljs"><code>' +
-          hljs.highlight(lang, str, true).value +
+          hljs.highlight(str, { language: lang, ignoreIllegals: true }).value +
           "</code></pre>"
         );
       } catch (__) { }